+++ frontend/packages/process-editor/src/components/Canvas/Canvas.tsx
@@ -40,7 +40,7 @@ export const Canvas = ({ onSave }: CanvasProps): JSX.Element => {
 
   useConfirmationDialogOnPageLeave(
     !!numberOfUnsavedChanges,
-    t('process_editor.unsaved_changes_warning'),
+    t('process_editor.unsaved_changes_confirmation_message'),
   );
 
   return (
